+++ src/clj_kondo/impl/analyzer/namespace.clj
@@ -369,6 +369,10 @@
         unused-namespace-disabled? (identical? :off (-> ctx :config :linters :unused-namespace :level))
         analyzed
         (map (fn [[require-kw libspecs]]
+               (when-not libspecs
+                 (findings/reg-finding!
+                  ctx (node->line (:filename ctx) require-kw :syntax
+                                  "Invalid require: no libs specified to load")))
                (for [libspec-expr libspecs
                      normalized-libspec-expr (normalize-libspec ctx nil libspec-expr unused-namespace-disabled?)
                      analyzed (analyze-libspec ctx ns-name require-kw normalized-libspec-expr)]
@@ -668,6 +672,10 @@
                                                                  utils/symbol-from-token)))
                                       (second children)))))
                            children)]
+    (when-not (seq children)
+      (findings/reg-finding!
+       ctx (node->line (:filename ctx) require-node :syntax
+                       "Invalid require: no libs specified to load")))
     (when (some-> children first sexpr empty-spec?)
       (findings/reg-finding!
        ctx